refactor: extract setup test helper
